+++ app/vmui/packages/vmui/src/utils/storage.ts
@@ -21,6 +21,7 @@ export type StorageKeys = "AUTOCOMPLETE"
   | "METRICS_QUERY_HISTORY"
   | "SERVER_URL"
   | "RAW_JSON_LIVE_VIEW"
+  | "LOGS_OVERRIDE_TIME"
   | DeprecatedStorageKeys;